SFDC_Integration/requirements.txt
simple-salesforce
python-dotenv